python/requirements.txt
torch==1.6.0
absl-py==0.10.0
aiocontextvars==0.2.2
apex @ git+git://github.com/NVIDIA/apex.git@700d6825e205732c1d6be511306ca4e595297070
appdirs==1.4.4
astunparse==1.6.3
audioread==2.1.8
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
ConfigArgParse==1.2.3
contextvars==2.4
cycler==0.10.0
decorator==4.4.2
Deprecated==1.2.10
Flask==1.1.2
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.8
google-auth==1.21.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
immutables==0.14
importlib-metadata==1.7.0
itsdangerous==2.0.0a1
-e git+https://github.com/bloriot97/javalang.git@37ac7f4d7e859a859a6914b733dac59e8c32282f#egg=javalang
Jinja2==3.0.0a1
joblib==0.16.0
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
librosa==0.8.0
llvmlite==0.32.1
loguru==0.5.2
Markdown==3.2.2
MarkupSafe==2.0.0a1
matplotlib==3.3.1
matplotlib-venn==0.11.5
munch==2.5.0
numba==0.43.0
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.42
OpenNMT-py==1.2.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
pooch==1.2.0
pretrainedmodels==0.7.4
protobuf==4.0.0rc2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyonmttok==1.19.0
pyparsing==2.4.7
pyrouge==0.1.3
python-dateutil==2.8.1
python-intervals==1.10.0.post1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.6
scikit-learn==0.23.2
scipy==1.4.1
seaborn==0.11.0
six==1.15.0
smmap==3.0.4
SoundFile==0.10.3.post1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
terminaltables==3.1.0
threadpoolctl==2.1.0
torchtext==0.4.0
torchvision==0.7.0
tqdm==4.48.2
urllib3==1.25.10
waitress==1.4.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
paramiko==2.7.2
scp==0.13.3
